Renamespace AssertionFactory
